# flockr-admin/src/main/resources/webroot/swagger/flockr-admin.yaml
openapi: 3.0.1
info:
  title: Flockr Admin API
  description: "Flockr Admin Service - Manage audiences, rules, and data connectors"
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Local development server
...
paths:
...
  /v1/connectors/types/{typeId}:
    get:
      tags:
      - Data Connectors
      summary: Get connector type by ID
      description: Retrieves all details of an onboarded source or sink connector
        type by typeId
      operationId: getConnectorTypeById
      parameters:
      - name: typeId
        in: path
        description: ID of the connector type
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        "404":
          description: Connector type not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "400":
          description: Bad Request due to invalid typeId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
  /v1/datasinks:
    get:
      tags:
      - Data Connectors
      summary: List data sinks
      description: Retrieves a paginated list of data sinks
      operationId: listSinks
      parameters:
      - name: pageSize
        in: query
        description: Number of items per page
        schema:
          minimum: 1
          type: integer
          format: int32
          default: 10
        example: 10
      - name: pageNum
        in: query
        description: Page number (0-indexed)
        schema:
          minimum: 0
          type: integer
          format: int32
          default: 0
        example: 0
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        "400":
          description: Bad Request due to invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
  /v1/datasources:
    get:
      tags:
      - Data Connectors
      summary: List data sources
      description: Retrieves a paginated list of data sources
      operationId: listSources
      parameters:
      - name: pageSize
        in: query
        description: Number of items per page
        schema:
          minimum: 1
          type: integer
          format: int32
          default: 10
        example: 10
      - name: pageNum
        in: query
        description: Page number (0-indexed)
        schema:
          minimum: 0
          type: integer
          format: int32
          default: 0
        example: 0
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        "400":
          description: Bad Request due to invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
  /v1/connectors/types:
    get:
      tags:
      - Data Connectors
      summary: List connector types
      description: Retrieves a list of available data connector types
      operationId: listTypes
      parameters:
      - name: kind
        in: query
        description: Kind of connector (SOURCE or SINK)
        schema:
          type: string
          default: SOURCE
        example: SOURCE
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
  /v1/connectors/types/onboard:
    post:
      tags:
      - Data Connectors
      summary: Onboard a connector type
      description: Onboards a new data connector type (SOURCE or SINK)
      operationId: onboardConnectorType
      parameters:
      - name: email
        in: header
        description: User email for authentication
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardConnectorTypeRequest'
      responses:
        "200":
          description: Connector type onboarded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        "400":
          description: Bad Request due to invalid/missing body params
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
  /v1/datasinks/onboard:
    post:
      tags:
      - Data Connectors
      summary: Onboard a data sink
      description: Onboards a new data sink connector
      operationId: onboardSink
      parameters:
      - name: email
        in: header
        description: User email for authentication
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardDataSinkRequest'
      responses:
        "200":
          description: Data sink onboarded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        "400":
          description: Bad Request due to invalid/missing body params
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
  /v1/datasources/onboard:
    post:
      tags:
      - Data Connectors
      summary: Onboard a data source
      description: Onboards a new data source connector
      operationId: onboardSource
      parameters:
      - name: email
        in: header
        description: User email for authentication
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardDataSourceRequest'
      responses:
        "200":
          description: Data source onboarded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        "400":
          description: Bad Request due to invalid/missing body params
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
components:
  schemas:
    Success:
      type: object
      properties:
        data:
          type: object
    ErrorEntity:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        cause:
          type: string
    Failure:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorEntity'
...
    JsonObject:
      type: object
      properties:
        map:
          type: object
          additionalProperties:
            type: object
        empty:
          type: boolean
...
    OnboardConnectorTypeRequest:
      required:
      - displayName
      - kind
      - type
      type: object
      properties:
        kind:
          type: string
        type:
          type: string
        displayName:
          type: string
        configSchema:
          $ref: '#/components/schemas/JsonObject'
...
    OnboardDataSinkRequest:
      required:
      - name
      - typeId
      type: object
      properties:
        name:
          type: string
        typeId:
          type: integer
          format: int64
        config:
          $ref: '#/components/schemas/JsonObject'
...
    OnboardDataSourceRequest:
      required:
      - name
      - typeId
      type: object
      properties:
        name:
          type: string
        typeId:
          type: integer
          format: int64
        config:
          $ref: '#/components/schemas/JsonObject'